+++ src/map/editor/state/clipboard/mod.rs
@@ -165,7 +165,12 @@ impl ClipboardData
 
                 if let Some(h) = data.sprite_hull(drawing_resources, grid)
                 {
-                    hull = hull.merged(&h);
+                    hull = hull.merged(
+                        &Hull::from_points(
+                            h.rectangle().into_iter().map(|vx| grid.point_projection(vx))
+                        )
+                        .unwrap()
+                    );
                 }
 
                 match data.path_hull()